Another attempt to fix win64 compile.
BUG=
TEST=

Review URL: http://codereview.chromium.org/6525051

git-svn-id: http://v8.googlecode.com/svn/branches/bleeding_edge@6818 ce2b1a6d-e550-0410-aec6-3dcde31c8c00

+++ test/cctest/test-parsing.cc
@@ -321,15 +321,16 @@ TEST(Regress928) {
 
   data->Initialize();
 
-  intptr_t first_function = strstr(program, "function") - program;
-  intptr_t first_lbrace = first_function + strlen("function () ");
+  int first_function =
+      static_cast<int>(strstr(program, "function") - program);
+  int first_lbrace = first_function + strlen("function () ");
   CHECK_EQ('{', program[first_lbrace]);
   i::FunctionEntry entry1 = data->GetFunctionEntry(first_lbrace);
   CHECK(!entry1.is_valid());
 
-  intptr_t second_function =
-      strstr(program + first_lbrace, "function") - program;
-  intptr_t second_lbrace = second_function + strlen("function () ");
+  int second_function =
+      static_cast<int>(strstr(program + first_lbrace, "function") - program);
+  int second_lbrace = second_function + strlen("function () ");
   CHECK_EQ('{', program[second_lbrace]);
   i::FunctionEntry entry2 = data->GetFunctionEntry(second_lbrace);
   CHECK(entry2.is_valid());
